Project ready to upload to server

+++ tienda_digital/shop/models.py
@@ -8,6 +8,6 @@ def upload_to(instance, filename):
 class Productos(models.Model):
     sku = models.CharField(max_length=30)
     name = models.CharField(max_length=50)
-    price = models.FloatField(default=0.0)
+    price = models.IntegerField(default=0)
     stock = models.IntegerField(default=0)
     image = models.ImageField(upload_to= upload_to, default="")
